- Andries Brouwer: final isofs pieces. - Kai Germaschewski: ISDN - play CD audio correctly, don't stop after 12 minutes. - Anton Altaparmakov: disable NTFS mmap for now, as it doesn't work. - Stephen Tweedie: fix inode dirty block handling - Bill Hartner: reschedule_idle - prefer right cpu - Johannes Erdfelt: USB updates - Alan Cox: synchronize - Richard Henderson: alpha updates and optimizations - Geert Uytterhoeven: fbdev could be fooled into crashing fix - Trond Myklebust: NFS filehandles in inode rather than dentry

+++ arch/i386/kernel/irq.c
@@ -179,7 +179,7 @@ int get_irq_list(char *buf)
 
 #ifdef CONFIG_SMP
 unsigned char global_irq_holder = NO_PROC_ID;
-unsigned volatile int global_irq_lock;
+unsigned volatile long global_irq_lock; /* pendantic: long for set_bit --RR */
 
 extern void show_stack(unsigned long* esp);
